requirements.txt
GitPython
PyGithub
dbt-core>=1.4,<1.5
dbt-duckdb
duckdb
piperider[duckdb]
